# routers/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for CodeCover",
        "title": "CodeCover API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/reports/{id}/repo": {
            "get": {
                "tags": [
                    "Report"
                ],
                "summary": "get repository of the report id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "report id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "repository",
                        "schema": {
                            "$ref": "#/definitions/core.Repo"
                        }
                    },
                    "400": {
                        "description": "error message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/repos": {
            "get": {
                "tags": [
                    "Repository"
                ],
                "summary": "List repositories for all available SCM providers",
                "responses": {
                    "200": {
                        "description": "repositories",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/core.Repo"
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Repository"
                ],
                "summary": "Create new repository for code coverage",
                "parameters": [
                    {
                        "description": "repository to create",
                        "name": "repo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/core.Repo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created repository",
                        "schema": {
                            "$ref": "#/definitions/core.Repo"
                        }
                    }
                }
            }
        },
        "/repos/{scm}": {
            "get": {
                "tags": [
                    "Repository"
                ],
                "summary": "List repositories",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SCM source (github, gitea)",
                        "name": "scm",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "repositories",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/core.Repo"
                            }
                        }
                    }
                }
            }
        },
        "/repos/{scm}/{namespace}/{name}": {
            "get": {
                "tags": [
                    "Repository"
                ],
                "summary": "get repository",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SCM",
                        "name": "scm",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Namespace",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/core.Repo"
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Repository"
                ],
                "summary": "sync repository information with SCM",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SCM",
                        "name": "scm",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Namespace",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/core.Repo"
                        }
                    }
                }
            }
        },
        "/repos/{scm}/{namespace}/{name}/branches": {
            "get": {
                "tags": [
                    "Repository"
                ],
                "summary": "list repository branches",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SCM",
                        "name": "scm",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Namespace",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/repos/{scm}/{namespace}/{name}/commits": {
            "get": {
                "tags": [
                    "Repository"
                ],
                "summary": "list recent commits",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SCM",
                        "name": "scm",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Namespace",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "branch to list commits from",
                        "name": "gitref",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/core.Commit"
                            }
                        }
                    }
                }
            }
        },
        "/repos/{scm}/{namespace}/{name}/content/{path}": {
            "get": {
                "tags": [
                    "Repository"
                ],
                "summary": "Get a file content",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SCM",
                        "name": "scm",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Namespace",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "file path",
                        "name": "path",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "specify git ref, default main branch",
                        "name": "gitref",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "file content",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/repos/{scm}/{namespace}/{name}/files": {
            "get": {
                "tags": [
                    "Repository"
                ],
                "summary": "List all files in repository",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SCM",
                        "name": "scm",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Namespace",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "specify git ref, default main branch",
                        "name": "gitref",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "files",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/repos/{scm}/{namespace}/{name}/report": {
            "patch": {
                "tags": [
                    "Repository"
                ],
                "summary": "renew repository report id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SCM",
                        "name": "scm",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Namespace",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "updated repository",
                        "schema": {
                            "$ref": "#/definitions/core.Repo"
                        }
                    }
                }
            }
        },
        "/repos/{scm}/{namespace}/{name}/setting": {
            "get": {
                "tags": [
                    "Repository"
                ],
                "summary": "get repository setting",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SCM",
                        "name": "scm",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Namespace",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/core.RepoSetting"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Repository"
                ],
                "summary": "update repository setting",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SCM",
                        "name": "scm",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Namespace",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "repository setting",
                        "name": "setting",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/core.RepoSetting"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/core.RepoSetting"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "core.Commit": {
            "type": "object",
            "properties": {
                "committer": {
                    "type": "string"
                },
                "committerAvatar": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "sha": {
                    "type": "string"
                }
            }
        },
        "core.Repo": {
            "type": "object",
            "properties": {
                "branch": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "nameSpace": {
                    "type": "string"
                },
                "private": {
                    "type": "boolean"
                },
                "reportID": {
                    "type": "string"
                },
                "scm": {
                    "type": "SCMProvider"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "core.RepoSetting": {
            "type": "object",
            "properties": {
                "filters": {
                    "type": "FileNameFilters"
                },
                "mergePR": {
                    "type": "boolean"
                },
                "protected": {
                    "type": "boolean"
                },
                "updateAction": {
                    "type": "ReportUpdateAction"
                }
            }
        },
...
    }
}